API/requirements.txt
Flask
requests
python-dotenv
uuid
